Prevent double-error message on rejected calls.
If no `session.initial_destinations` is provided it is up to the code that rejected the call to send back an error message.

Also, not having `initial_destinatins` is a normal occurrence of a rejected call, not an error worthy of alerting.

Finally, do not trip FreeSwitch/Sofia, skip processing of an empty list of destinations.