queued call notifications should refer to number-domain
The `domain` in black-metal/call clearly refers to the centrex' domain,
not the DNS domain. Therefor it makes more sense to notify using
`number_domain:` in the key, similarly to what is done in middleware/setup
`report()` and `notify()`.

Of course the fact that for Centrex the decision was made to use the same
value for number-domain, DNS-domain, and Centrex-domain, only adds to the
confusion.

Keys starting with `domain:` should be reserved to things related to the
DNS name, such as `domain:` records, endpoints, etc.

(This is confusing because the code also uses `set_domain` for Reference
objects, where this clearly refers to a DNS-domain used by endpoints.
Also, the `get_domain()`/`set_domain()` pair in black-metal/call is somewhat
misleading, although the intent is clearly to refer to a Centrex domain.)